:white_check_mark: Add templatetags tests
